feat: rename ToggleSelector
